+++ crates/backend/executor/nodeagg/src/gsets.rs
@@ -65,7 +65,10 @@ pub(crate) struct PerHashSetData<'mcx> {
     largest_grp_col_idx: i32,
     grouped_cols: PgVec<'mcx, i16>,
     cell: NonNull<NonNull<AggPerGroup>>,
-    hashiter: usize,
+    // u64: in hashtable mode this is execgrouping's packed (start,visited)
+    // cursor, whose high-32 packing must survive 32-bit wasm; in compact
+    // mode it is a plain row index (cast at use).
+    hashiter: u64,
     // hash_agg_set_limits' input_groups for this set's hashagg_spill_init
     // calls (C: perhash->aggnode->numGroups).
     num_groups: f64,
